Refactor the fallback mounter exception to be useable by api consumers
